fix(object-store): drop three redundant u64 -> u64 casts
`clippy::unnecessary_cast` at src/framework/object_store.rs:888, 891 and 901
(`meta.size as u64` where `meta.size` is already `u64`), failing the
"Clippy + test the cloud feature" gate on the MSRV.

Pre-existing, and hidden the same way the p2p lint was: this code sits in the
object-store adapter compiled only behind the `cloud` feature, and every
verification I ran was `--workspace` with DEFAULT features, which never
compiles it. Two gated feature trees, two lint sets nobody was running locally.

Now checked the way the workflow actually checks: each of default, `--features
p2p`, `--features cloud`, and `--workspace` linted from a FRESH pass (`touch
src/lib.rs` first — clippy's incremental cache will otherwise replay a
diagnostic against source that has already been corrected, which cost a
round-trip earlier in this sequence).

Verified on the gating toolchain (1.88.0): all four clippy invocations exit 0;
`test --features cloud --lib cloud` green; workspace tests green.

AI-Assisted-By: {"agent":"claude-code","id":"ai:anthropic-claude-code-claude-code","model":"claude-opus-5","product":"Claude Code","provider":"Anthropic","roles":["author","tester"]}
File-Contribution: {"by":[{"actor":"ai:anthropic-claude-code-claude-code","role":"author"}],"operation":"modify","path":"src/framework/object_store.rs","summary":"remove three no-op u64 casts flagged under the cloud feature"}

+++ src/framework/object_store.rs
@@ -885,10 +885,10 @@ async fn read_provider_capped(
         Err(object_store::Error::NotFound { .. }) => return Ok(None),
         Err(error) => return Err(StoreError::backend("head", key, error)),
     };
-    if meta.size as u64 > MAX_STORED_SIZE {
+    if meta.size > MAX_STORED_SIZE {
         return Err(StoreError::TooLarge {
             key,
-            size: (meta.size as u64).saturating_sub(OBJECT_HEADER_LEN as u64),
+            size: meta.size.saturating_sub(OBJECT_HEADER_LEN as u64),
             max: MAX_OBJECT_SIZE,
         });
     }
@@ -898,7 +898,7 @@ async fn read_provider_capped(
         Err(error) => return Err(StoreError::backend("get", key, error)),
     };
     let mut stream = result.into_stream();
-    let mut stored = Vec::with_capacity((meta.size as u64).min(8 * 1024 * 1024) as usize);
+    let mut stored = Vec::with_capacity(meta.size.min(8 * 1024 * 1024) as usize);
     while let Some(chunk) = stream.next().await {
         let chunk = chunk.map_err(|error| StoreError::backend("read", key, error))?;
         let next_len = stored
